requirements.txt
python-telegram-bot==22.5
aiohttp==3.9.5
aiofiles==24.1.0
openpyxl==3.1.5
psycopg2-binary==2.9.9
python-dotenv==1.0.1

